Refactor question component to allow answer input from outside. Behaviour previously: AbstractQuestionComponent use the route parameters to get the current questions.
Behaviour now:
AbstractQuestionComponent do not specify how to get the answers of the
user. It now uses an abstract method to get the answers. The sub
component QuestionComponent use the same behaviour as before. However,
QuestionOverviewComponent gets the questions from outside, so in the
ResultComponent it is now possible to specify different or
pre-calculated Answers.