Refactor question component and restructure observable structure.
Behaviour previously:
Mixed input and output structure for abstractQuestionComponent,
partially also direct access to url parameters.

Behaviour now:
Clean structure of input and output for questions. The conversion of the
url parameter is now handled by the AbstractQuestionComponent. The
url parameter is handled by currentAnswers (input) and newAnswers
(output). Converted result is with all information in
questionWithUserAnswers. Those can be handled in the component or
published to the called component. This should not be used for url
handling.