Feature: result directly after questions
Behavior before: after answering all the questions the user gets
redirected to the location page, where he/she is able to select his
location. The user is also able to skip this step.

Behavior after: the user is directly redirect to the result page with
a maximized map. The user is able to see the same dialog as before to
select his address to keep him focused on changing the address.