cleanup dependencies and explicit terminate import app
The import application itself is just not terminating by default. The reason
therefore is that the elasticsearch rest client is keeping the thread alive.
Since elasticsearch is mandatory for importing this can't be removed.
However, the behavior with that is as expected.
See https://stackoverflow.com/questions/48328226 for more information.